trying to return a copy of the view and delete the whole data structure when using a static plotting backend to limit the amount of memory used in long docs notebooks
